[ch135] Refine code and optimize search icon
